Refactor top level components
Successfully connected top level components to the ColorFamilyView
