+++ components/index.js
@@ -1,8 +1,36 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import $ from 'jquery';
 import ColorFamilyView from './ColorFamilyView.js';
 
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      currentFilter: 'mostClicked',
+      colorFamilies: []
+    };
+  }
+
+  componentWillMount() {
+    $.ajax({
+      url: '/api/colors',
+      success: function(data) {
+        this.setState({ colorFamilies: data });
+        console.log(this.state.colorFamilies);
+      }.bind(this),
+      dataType: 'JSON'
+    });
+  }
+
+  render() {
+    return (
+      <ColorFamilyView colorFamilies={this.state.colorFamilies}/>
+    );
+  }
+}
+
 ReactDOM.render(
-  <ColorFamilyView />,
+  <App />,
   document.getElementById('root')
 );
